move legacy gas calc file
