+++ tests/test_contract.py
@@ -5,7 +5,7 @@ import pytest
 import web3
 from web3 import Web3
 
-from telliot_core.contract.gas import fetch_gas_price
+from telliot_core.gas.legacy_gas import fetch_gas_price
 from telliot_core.queries.legacy_query import LegacyRequest
